+++ R/as-tsibble.R
@@ -24,21 +24,21 @@ globalVariables(c("key", "value", "zzz"))
 #'
 #' @examples
 #' # create a tsibble w/o a key ----
-#' tbl1 <- tsibble(
+#' tsbl1 <- tsibble(
 #'   date = seq(as.Date("2017-01-01"), as.Date("2017-01-10"), by = 1),
 #'   value = rnorm(10),
 #'   key = id(), index = date
 #' )
-#' tbl1
+#' tsbl1
 #'
 #' # create a tsibble with one key ----
-#' tbl2 <- tsibble(
+#' tsbl2 <- tsibble(
 #'   qtr = rep(yearquarter(seq(2010, 2012.25, by = 1 / 4)), 3),
 #'   group = rep(c("x", "y", "z"), each = 10),
 #'   value = rnorm(30),
 #'   key = id(group), index = qtr
 #' )
-#' tbl2
+#' tsbl2
 #'
 #' @export
 tsibble <- function(..., key = id(), index, regular = TRUE) {